Move SavedQueries to the Keen:: namespace
SavedQueries did not follow naming conventions elsewhere in the gem, and could cause name collisions. It has been moved out of the Client namespace into the base Keen namespace, and properly wrapped in the Keen module
