+++ lib/keen/client.rb
@@ -3,7 +3,6 @@ require 'keen/version'
 require 'keen/client/publishing_methods'
 require 'keen/client/querying_methods'
 require 'keen/client/maintenance_methods'
-require 'keen/client/saved_queries'
 require 'openssl'
 require 'multi_json'
 require 'base64'
